fixed lambdas
Lambdas in a list comprehension that use the for loop variable actually
use the variable by reference. The effect is that all the lambda
functions generated in the comprehension use the last value the variable
held. To combat this, you can wrap the lambda in a second lambda that
takes the for loop variable as an argument and explicitly call it in the
comprehension.

from this StackOverflow answer:
https://stackoverflow.com/a/34021333/4765406